feat(homeScreen): move ContributionGraph into its own home section
For hours-mode publishers (pioneer / aux / CO / special / custom), the
contribution graph is now a top-level Home section instead of being
buried inside the ProfileDetailOverlay. Visibility + ordering are
wired through the existing home-screen-elements preference so it
participates in the reorder UI alongside Timer, Service Report, etc.
A new `useDailyMinutes` hook hoists the cache-aware
`flattenDailyMinutes` logic that previously lived inline in the
overlay so both screens share the same flatten/cache.

Regular publishers (checkbox mode) see no change — the section is
hidden and the prefs row doesn't surface, matching how Timer is gated.
